requirements_ops.txt
virtualenv==15.1.0
fabric3==1.13.1
ansible==2.3.1.0
gunicorn==19.7.1
uwsgi==2.0.15




